Refactor settings access via SettingsManager
